+++ src/runtime/tooling/index.js
@@ -1,8 +1,8 @@
+import '../../intl/polyfill';
+
 import FTLASTParser from '../../ftl/ast/parser';
 import FTLEntriesParser from '../../ftl/entries/parser';
 import { createEntriesFromAST } from '../../ftl/entries/transformer';
-import { MessageContext } from '../../intl/context';
-import { FTLNumber, FTLDateTime } from '../../intl/types';
 
 import { L10nError } from '../../lib/errors';
 import { keysFromContext, valueFromContext, entityFromContext }
@@ -11,7 +11,6 @@ import { keysFromContext, valueFromContext, entityFromContext }
 import { fetchResource } from '../web/io';
 
 export default {
-  FTLASTParser, FTLEntriesParser, createEntriesFromAST,
-  MessageContext, FTLNumber, FTLBase, L10nError,
+  FTLASTParser, FTLEntriesParser, createEntriesFromAST, L10nError,
   keysFromContext, valueFromContext, entityFromContext, fetchResource
 };
